serverless/bucket_logger/requirements.txt
functions-framework==3.*
mysql-connector-python==8.*
google-cloud-storage==2.*